% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_ui.R
\docType{data}
\name{colBoxplot4}
\alias{colBoxplot4}
\title{User Interface of the elaborator application}
\format{
An object of class \code{character} of length 1.
}
\usage{
colBoxplot4
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
No return value. User interface part of the app, used in launch_elaborator-function.
}
\description{
User Interface of the elaborator application
}
\keyword{internal}
