% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator-package.R
\docType{package}
\name{elaborator-package}
\alias{elaborator}
\alias{elaborator-package}
\title{elaborator: A 'shiny' Application for Exploring Laboratory Data}
\description{
A novel concept for generating knowledge and gaining insights into laboratory data. You will be able to efficiently and easily explore your laboratory data from different perspectives. Janitza, S., Majumder, M., Mendolia, F., Jeske, S., & Kulmann, H. (2021) \doi{10.1007/s43441-021-00318-4}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/elaborator}
  \item Report bugs at \url{https://github.com/openpharma/elaborator/issues}
}

}
\author{
\strong{Maintainer}: Bodo Kirsch \email{bodo.kirsch@bayer.com}

Authors:
\itemize{
  \item Silke Janitza
  \item Steffen Jeske
  \item Madhurima Majumder
  \item Franco Mendolia
  \item Hermann Kulmann
}

}
\keyword{internal}
