% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_calculate_color_index.R
\name{elaborator_calculate_color_index}
\alias{elaborator_calculate_color_index}
\title{elaborator_calculate_color_index - calculates the index of the color vector in the qualitative trends plot}
\usage{
elaborator_calculate_color_index(r)
}
\arguments{
\item{r}{numeric value}
}
\description{
elaborator_calculate_color_index - calculates the index of the color vector in the qualitative trends plot
}
\keyword{internal}
