% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_calculate_pattern_number.R
\name{elaborator_calculate_pattern_number}
\alias{elaborator_calculate_pattern_number}
\title{elaborator_calculate_pattern_number - adds one to the pattern for a ternary representation}
\usage{
elaborator_calculate_pattern_number(p)
}
\arguments{
\item{p}{pattern}
}
\description{
elaborator_calculate_pattern_number - adds one to the pattern for a ternary representation
}
\keyword{internal}
