% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_calculate_ref_pattern.R
\name{elaborator_calculate_ref_pattern}
\alias{elaborator_calculate_ref_pattern}
\title{elaborator_calculate_ref_pattern - returns a specific row of a combination matrix}
\usage{
elaborator_calculate_ref_pattern(index, number_combinations)
}
\arguments{
\item{index}{row of combination matrix}

\item{number_combinations}{length of the combination vector}
}
\description{
elaborator_calculate_ref_pattern - returns a specific row of a combination matrix
}
\keyword{internal}
