% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_calculate_spearman_distance.R
\name{elaborator_calculate_spearman_distance}
\alias{elaborator_calculate_spearman_distance}
\title{elaborator_calculate_spearman_distance - function to calculate distance matrix with spearman method}
\usage{
elaborator_calculate_spearman_distance(x)
}
\arguments{
\item{x}{data matrix for which the spearman distance should be calculated}
}
\description{
elaborator_calculate_spearman_distance - function to calculate distance matrix with spearman method
}
\keyword{internal}
