% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_calculate_test_for_all_visits.R
\name{elaborator_calculate_test_for_all_visits}
\alias{elaborator_calculate_test_for_all_visits}
\title{Perform pairwise statistical tests for quantitative trend plots}
\usage{
elaborator_calculate_test_for_all_visits(
  elab_data,
  Visit1,
  Visit2,
  sign_test,
  pcutoff
)
}
\arguments{
\item{elab_data}{nested elaborator data with possibly up to three nests 'raw',}

\item{Visit1}{first visit for statistical tests (only one allowed)}

\item{Visit2}{second visit(s) for statistical tests (multiple visits applicable - pairwise tests performed)}

\item{sign_test}{character string if sign test or t-test should be performed ("signtest"/"ttest")}

\item{pcutoff}{p value for test(s)}
}
\value{
tibble with test results (estimate/p-values) and color code for quantitative trend graphs
}
\description{
Perform pairwise statistical tests for quantitative trend plots
}
\keyword{internal}
