% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_change_class_required_variables.R
\name{elaborator_change_class_required_variables}
\alias{elaborator_change_class_required_variables}
\title{function to change (rdata) or set (csv data) the classes of the required variables in
elaborator}
\usage{
elaborator_change_class_required_variables(elab_data, visit, treatment, lab)
}
\arguments{
\item{elab_data}{elaborator_data (requires columns AVISIT,LBTESTCD,TRTP,SUBJIDN,LBORRES,LBORNRHI,LBORNRLO)}

\item{visit}{desired level order for AVISIT}

\item{treatment}{desired level order for TRTP}

\item{lab}{desired level order for LBTESTCD}
}
\value{
elab_data with changed classes (and factor levels)
}
\description{
function to change (rdata) or set (csv data) the classes of the required variables in
elaborator
}
\keyword{internal}
