% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_create_hover_info_text.R
\name{elaborator_create_hover_info_text}
\alias{elaborator_create_hover_info_text}
\title{create hover text info in elaborator app zoom panel(s)}
\usage{
elaborator_create_hover_info_text(elab_data, labparameter, treat, select.visit)
}
\arguments{
\item{elab_data}{elaborator data}

\item{labparameter}{labparameter used for filtering}

\item{treat}{treatment used for filtering}
}
\value{
character string with text about number of visits in every
        treatment group and labparameter
}
\description{
create hover text info in elaborator app zoom panel(s)
}
\keyword{internal}
