% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_demo_data.R
\name{elaborator_demo_rdata_available}
\alias{elaborator_demo_rdata_available}
\title{Whether demo `.RData` is present at the installed package path.}
\usage{
elaborator_demo_rdata_available()
}
\description{
Whether demo `.RData` is present at the installed package path.
}
\keyword{internal}
