% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_derive_equal_values.R
\name{elaborator_derive_equal_values}
\alias{elaborator_derive_equal_values}
\title{Returns indices for all elements of a vector x that are contained in another vector y}
\usage{
elaborator_derive_equal_values(y, x)
}
\arguments{
\item{y}{numeric vector}

\item{x}{numeric vector}
}
\value{
A numeric vector with the indices for all elements of a vector x that are contained in another vector y.
}
\description{
This function is mostly useful for generating the qualitative trend analysis plots. It returns the indices or all elements of a vector x that are contained in another vector y
}
\keyword{internal}
