% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_derive_test_values.R
\name{elaborator_derive_test_values}
\alias{elaborator_derive_test_values}
\title{elaborator_derive_test_values - derives p-values/estimates for sign-test or t-test (by applying functions specified above)}
\usage{
elaborator_derive_test_values(
  data = data,
  signtest = TRUE,
  Visit1 = c("Randomization"),
  Visit2 = c("End of Treatment"),
  lab_column
)
}
\arguments{
\item{data}{data set}

\item{signtest}{boolean value if signtest should be performded (TRUE/FALSE)}

\item{Visit1}{visit time point one}

\item{Visit2}{visit time point two}

\item{lab_column}{lab parameter column}
}
\description{
elaborator_derive_test_values - derives p-values/estimates for sign-test or t-test (by applying functions specified above)
}
\keyword{internal}
