% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_draw_boxplot_color.R
\name{elaborator_draw_boxplot_color}
\alias{elaborator_draw_boxplot_color}
\title{elaborator_draw_boxplot_color - Color preview function}
\usage{
elaborator_draw_boxplot_color(x)
}
\arguments{
\item{x}{color palette}
}
\description{
elaborator_draw_boxplot_color - Color preview function
}
\keyword{internal}
