% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_draw_curved_line.R
\name{elaborator_draw_curved_line}
\alias{elaborator_draw_curved_line}
\title{elaborator_draw_curved_line - Draw a curved line from start to end point}
\usage{
elaborator_draw_curved_line(x1, y1, x2, y2, ...)
}
\arguments{
\item{x1}{x-coordinate of the starting point}

\item{y1}{y-coordinate of the starting point}

\item{x2}{x-coordinate of the end point}

\item{y2}{y-coordinate of the end point}

\item{...}{further parameter}
}
\description{
elaborator_draw_curved_line - Draw a curved line from start to end point
}
\keyword{internal}
