% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_draw_circle.R,
%   R/elaborator_draw_dots.R
\name{elaborator_draw_dots}
\alias{elaborator_draw_dots}
\title{elaborator_draw_dots - Draw a circle in a plot window with specific size/coordinates and colors}
\usage{
elaborator_draw_dots(
  x,
  y,
  height,
  dot_col,
  pattern_Matrix,
  number_column,
  dot_Radius,
  dot_radius,
  fontsize,
  empty_color = "#A9A9A9",
  upper_color = "#2fb39f",
  lower_color = "#f78300"
)

elaborator_draw_dots(
  x,
  y,
  height,
  dot_col,
  pattern_Matrix,
  number_column,
  dot_Radius,
  dot_radius,
  fontsize,
  empty_color = "#A9A9A9",
  upper_color = "#2fb39f",
  lower_color = "#f78300"
)
}
\arguments{
\item{x}{x-coordinate of total dot}

\item{y}{y-coordinate of total dot}

\item{height}{height of curved Lines}

\item{dot_col}{color of total dot}

\item{pattern_Matrix}{Matrix with the Reference-value based pattern information}

\item{number_column}{number of columns (visits)}

\item{dot_Radius}{dot radius}

\item{dot_radius}{empty group dot radius}

\item{fontsize}{font size}
}
\description{
elaborator_draw_dots - Draw a circle in a plot window with specific size/coordinates and colors

elaborator_draw_dots - Draw a circle in a plot window with specific size/coordinates and colors
}
\keyword{elaborator_curvedLine}
\keyword{internal}
\keyword{required}
