% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_draw_ref_pattern.R
\name{elaborator_draw_ref_pattern}
\alias{elaborator_draw_ref_pattern}
\title{elaborator_draw_ref_pattern - create the Reference-value based pattern plots}
\usage{
elaborator_draw_ref_pattern(
  pattern_Matrix,
  fontsize,
  numberColumns,
  ColorBG = "#E2F3F2"
)
}
\arguments{
\item{pattern_Matrix}{matrix with reference-value based pattern information}

\item{fontsize}{font size}

\item{numberColumns}{number of columns}

\item{ColorBG}{background color}
}
\description{
elaborator_draw_ref_pattern - create the Reference-value based pattern plots
}
\keyword{internal}
