% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_draw_scheme_preview.R
\name{elaborator_draw_scheme_preview}
\alias{elaborator_draw_scheme_preview}
\title{Color scheme preview}
\usage{
elaborator_draw_scheme_preview(x)
}
\arguments{
\item{x}{color palette}
}
\value{
No return value; this function is used to generate plots with the selected color scheme for the qualitative trend analysis.
}
\description{
This function is mostly useful for changing the color scheme of the qualitative trend analysis plots.
}
\keyword{internal}
