% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_fill_with_missings.R
\name{elaborator_fill_with_missings}
\alias{elaborator_fill_with_missings}
\title{Transform data and fill them with missing values in app elaborator}
\usage{
elaborator_fill_with_missings(elab_data)
}
\arguments{
\item{elab_data}{data set}
}
\value{
Transformed data frame
}
\description{
Transform data and fill them with missing values in app elaborator
}
\keyword{internal}
