% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_load_and_check.R
\name{elaborator_load_and_check}
\alias{elaborator_load_and_check}
\title{loading function for elaborator data, also performs checks for required variables}
\usage{
elaborator_load_and_check(
  data_switch = "*.RData file",
  rdata_file_path = NULL,
  csv_file_path = NULL,
  loaded_file = NULL,
  separator = NULL,
  quote = NULL,
  decimal = NULL
)
}
\arguments{
\item{data_switch}{app widgets input wheter rdata or csv files or demo data should be uploaded}

\item{rdata_file_path}{path of rdata file}

\item{csv_file_path}{path of csv file}

\item{loaded_file}{saved file for demo data (outdated)}

\item{separator}{separator (for csv file upload only)}

\item{quote}{quote (for csv file upload only)}

\item{decimal}{decimal (for csv file upload only)}
}
\value{
list with data and error message
}
\description{
loading function for elaborator data, also performs checks for required variables
}
\keyword{internal}
