% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_perform_binomial_test.R
\name{elaborator_perform_binomial_test}
\alias{elaborator_perform_binomial_test}
\title{elaborator_perform_binomial_test - calculates binomial test for the change of a labparameter from one visit to another}
\usage{
elaborator_perform_binomial_test(
  data,
  treatment,
  lab_parameter,
  Visit1 = "Randomization",
  Visit2 = "End of Treatment",
  lab_column
)
}
\arguments{
\item{data}{data set}

\item{treatment}{treatment variable}

\item{lab_parameter}{llab parameter variable of the data set}

\item{Visit1}{visit time point one}

\item{Visit2}{visit time point two}

\item{lab_column}{lab parameter column}
}
\description{
elaborator_perform_binomial_test - calculates binomial test for the change of a labparameter from one visit to another
}
\keyword{internal}
