% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_perform_ttest.R
\name{elaborator_perform_ttest}
\alias{elaborator_perform_ttest}
\title{elaborator_perform_ttest - performs t-test for changes of a labparameter from one visit to another}
\usage{
elaborator_perform_ttest(
  data,
  treatment,
  lab_parameter,
  Visit1 = "Randomization",
  Visit2 = "End of Treatment",
  lab_column
)
}
\arguments{
\item{data}{data set}

\item{treatment}{treatment variable of the data set}

\item{lab_parameter}{lab parameter variable of the data set}

\item{lab_column}{lab parameter column}

\item{visit1}{visit Time point one}

\item{visit2}{visit Time point two}
}
\description{
elaborator_perform_ttest - performs t-test for changes of a labparameter from one visit to another
}
\keyword{internal}
