% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_perform_ttest2.R
\name{elaborator_perform_ttest2}
\alias{elaborator_perform_ttest2}
\title{Performs t-test for changes of a laboratory parameter between two visits}
\usage{
elaborator_perform_ttest2(
  data,
  treatment,
  lab_parameter,
  Visit1 = "Randomization",
  Visit2 = "End of Treatment",
  lab_column
)
}
\arguments{
\item{data}{data set}

\item{treatment}{name of treatment group}

\item{lab_parameter}{name of laboratory parameter}

\item{Visit1}{name of visit time point one}

\item{Visit2}{name of visit time point two}

\item{lab_column}{column name of data which includes the laboratory parameter names}
}
\value{
A list of class "htest" with the results of the t-test.
}
\description{
This function is mostly useful for generating the quantitative trend analysis plots. It returns the output of a t-test result used for colorizing the background of the plots.
}
\keyword{internal}
