% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_perform_unlist.R
\name{elaborator_perform_unlist}
\alias{elaborator_perform_unlist}
\title{elaborator_perform_unlist - function to unlist}
\usage{
elaborator_perform_unlist(y, x)
}
\arguments{
\item{y}{number of subjects}

\item{x}{number of treatments}
}
\description{
elaborator_perform_unlist - function to unlist
}
\keyword{internal}
