% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_plot_qual_trends.R
\name{elaborator_plot_qual_trends}
\alias{elaborator_plot_qual_trends}
\title{elaborator_plot_qual_trends - creates the qualitative trends plot in the shiny app elaborator}
\usage{
elaborator_plot_qual_trends(
  dat1,
  Variab,
  fontsize = 1,
  percent = 0.1,
  method = "InQuRa",
  color_palette = c("#ffffff", "#ffffcc", "#ffeda0", "#fed976", "#feb24c", "#fd8d3c",
    "#fc4e2a", "#e31a1c", "#bd0026", "#800026", "#000000"),
  Summa = Summa
)
}
\arguments{
\item{dat1}{data set}

\item{Variab}{variable}

\item{fontsize}{font size}

\item{percent}{percentage}

\item{method}{method}

\item{color_palette}{color vector}

\item{Summa}{summarization matrix}
}
\description{
elaborator_plot_qual_trends - creates the qualitative trends plot in the shiny app elaborator
}
\keyword{internal}
