% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_plot_ref_pattern.R
\name{elaborator_plot_ref_pattern}
\alias{elaborator_plot_ref_pattern}
\title{elaborator_plot_ref_pattern - create reference-value based pattern plots (only works in shiny applications)}
\usage{
elaborator_plot_ref_pattern(
  data,
  criterion,
  fontsize = 0.6,
  sorting_vector = as.character(unique(data[, "PARAMCD"]))
)
}
\arguments{
\item{data}{data set}

\item{criterion}{criterion}

\item{fontsize}{font size}

\item{sorting_vector}{sorting vector}
}
\description{
elaborator_plot_ref_pattern - create reference-value based pattern plots (only works in shiny applications)
}
\keyword{internal}
