% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_prepare_clustering_matrix.R
\name{elaborator_prepare_clustering_matrix}
\alias{elaborator_prepare_clustering_matrix}
\title{Transform data in form for cluster analysis in app elaborator}
\usage{
elaborator_prepare_clustering_matrix(elab_data, first_variable, last_variable)
}
\arguments{
\item{elab_data}{data set}

\item{first_variable}{character with variable name}

\item{last_variable}{character with variable name}
}
\value{
Transformed data frame
}
\description{
Transform data in form for cluster analysis in app elaborator
}
\keyword{internal}
