% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/elaborator_remove_visits_due_tolerated_missings.R
\name{elaborator_remove_visits_due_tolerated_missings}
\alias{elaborator_remove_visits_due_tolerated_missings}
\title{Transform data by removing visits with too many missing values in app elaborator}
\usage{
elaborator_remove_visits_due_tolerated_missings(elab_data, tolerated_value)
}
\arguments{
\item{elab_data}{data set}

\item{tolerated_value}{numeric value with tolerated percentage value}
}
\value{
Transformed data frame
}
\description{
Transform data by removing visits with too many missing values in app elaborator
}
\keyword{internal}
