% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_server.R
\name{elaborator_server}
\alias{elaborator_server}
\title{elaborator_server - Server Part of the elaborator application}
\usage{
elaborator_server(input, output, session)
}
\description{
elaborator_server - Server Part of the elaborator application
}
\keyword{internal}
