% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_tidy_data.R
\name{elaborator_tidy_data}
\alias{elaborator_tidy_data}
\title{old version (needs to be removed)}
\usage{
elaborator_tidy_data(elab_data, tolerated_value)
}
\description{
old version (needs to be removed)
}
