% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_transform_transparent.R
\name{elaborator_transform_transparent}
\alias{elaborator_transform_transparent}
\title{Calculates the transparent version of a given color}
\usage{
elaborator_transform_transparent(some_color, alpha = 100)
}
\arguments{
\item{some_color}{A color in hexcode}

\item{alpha}{shading parameter}
}
\value{
A character string representing a color hexcode with added transparency.
}
\description{
Calculates the transparent version of a given color
}
\keyword{internal}
