% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_ui.R
\name{elaborator_ui}
\alias{elaborator_ui}
\title{User Interface of the elaborator application}
\usage{
elaborator_ui()
}
\value{
No return value. User interface part of the app, used in launch_elaborator-function.
}
\description{
User Interface of the elaborator application
}
\keyword{internal}
