% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_ui.R
\docType{data}
\name{jscode}
\alias{jscode}
\title{elaborator_ui - User Interface of the elaborator application}
\format{An object of class \code{character} of length 1.}
\usage{
jscode
}
\description{
elaborator_ui - User Interface of the elaborator application
}
\keyword{internal}
