% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_calculate_color_index.R
\name{elaborator_calculate_color_index}
\alias{elaborator_calculate_color_index}
\title{Calculates the index of the color vector in the qualitative trend analysis plot}
\usage{
elaborator_calculate_color_index(r)
}
\arguments{
\item{r}{numeric value specifying the percentage of patients with a specific pattern of decreased/stable/increased values from one visit to the next}
}
\value{
A numeric value between 1 and 11. This value is used as an index for the color legend in the qualitative trends plot in which darker colors reflect more frequent patterns
}
\description{
This function is mostly useful for generating the qualitative trend analysis plots. It returns the index of the color for a percentage which is observed for a specific pattern characterized by decreases/stability/increases from one visit to the next.
}
\keyword{internal}
