% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_calculate_ref_pattern.R
\name{elaborator_calculate_ref_pattern}
\alias{elaborator_calculate_ref_pattern}
\title{Returns row of a matrix containing all combinations of elements 0 and 1 of a pre-specified length}
\usage{
elaborator_calculate_ref_pattern(index, number_combinations)
}
\arguments{
\item{index}{selected row -1 of the combination matrix which shall be returned}

\item{number_combinations}{length of the combination vector}
}
\value{
A numeric vector containing a specific combination of elements 0 and 1; this combination is used to calculate pattern for reference-value based pattern analysis plot.
}
\description{
This function is mostly useful for generating the reference-value based pattern analysis plots. It returns the (i-1)-th row of a matrix containing all combinations of elements 0 and 1 of a pre-specified length.
}
\keyword{internal}
