% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_calculate_spearman_distance.R
\name{elaborator_calculate_spearman_distance}
\alias{elaborator_calculate_spearman_distance}
\title{Calculates the distance matrix based on Spearman's correlation coefficient}
\usage{
elaborator_calculate_spearman_distance(x)
}
\arguments{
\item{x}{data matrix for which a distance matrix based on the Spearman correlation coefficient is calculated}
}
\value{
A distance matrix of class "dist", which can be used for example for a seriation algorithm
}
\description{
This function derives the Spearman correlation coefficients for a data matrix and transforms the resulting similarity matrix into a distance/dissimilarity matrix. No difference is made between positive and negative correlations, i.e., correlations of -1 and 1 will be transformed to a distance of 0; correlations of 0 will be transformed to a distance of 1. Missing values in the input data will be handled by casewise deletion.
}
\keyword{internal}
