% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_draw_curved_line.R
\name{elaborator_draw_curved_line}
\alias{elaborator_draw_curved_line}
\title{Draws a curved line from start to end point}
\usage{
elaborator_draw_curved_line(x1, y1, x2, y2, ...)
}
\arguments{
\item{x1}{x-coordinate of the starting point}

\item{y1}{y-coordinate of the starting point}

\item{x2}{x-coordinate of the end point}

\item{y2}{y-coordinate of the end point}

\item{...}{further parameters}
}
\value{
No return value; this function is called to draw curved lines of the reference-value based pattern plot.
}
\description{
This function is mostly useful for generating the reference-value based pattern analysis plots. It draws a curved line from a start to an end point.
}
\keyword{internal}
