% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_draw_dots.R
\name{elaborator_draw_dots}
\alias{elaborator_draw_dots}
\title{Draws a circle with specified size, coordinates and color in a plot window}
\usage{
elaborator_draw_dots(
  x,
  y,
  height,
  dot_col,
  pattern_Matrix,
  number_column,
  dot_Radius,
  dot_radius,
  fontsize,
  empty_color = "#A9A9A9",
  upper_color = "#2fb39f",
  lower_color = "#f78300"
)
}
\arguments{
\item{x}{x-coordinate of circle}

\item{y}{y-coordinate of circle}

\item{height}{height of curved lines which connect circles}

\item{dot_col}{color of circle}

\item{pattern_Matrix}{matrix with the reference-value based pattern information}

\item{number_column}{number of layers (visits)}

\item{dot_Radius}{circle radius for circles representing existing patterns}

\item{dot_radius}{circle radius for circles representing non-existing patterns}

\item{fontsize}{font size of numbers printed inside the circles}
}
\value{
No return value; this function is called to draw the circles for the reference-value based pattern analysis plot.
}
\description{
This function is mostly useful for generating the reference-value based pattern analysis plots. It draws a circle with specified size, coordinates and color in a plot window.
}
\keyword{internal}
