% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elaborator_server.R
\name{elaborator_server}
\alias{elaborator_server}
\title{Server part of the elaborator application}
\usage{
elaborator_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for shiny.}
}
\value{
No return value. Server part of the app, used in launch_elaborator-function.
}
\description{
Server part of the elaborator application
}
\keyword{internal}
